+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial implementation of Delete command.
 - Add functions "cache_input", "blob_from_file".
 - Extended support for BLOBs.
+- Allow to set the list of allowed commands to execute.
 
 ### Changed